// src/app/shared/models/environmental-stats.interface.ts
export interface IEnvironmentalStats {
    totalRenewableElectricityConsumed: number;
    totalElectricityConsumed: number;
    totalConsumedGas: number;
    totalConsumedWater: number;
    totalDangerousWasteGenerated: number;
    totalNonDangerousWasteGenerated: number;
    totalWasteSold: number;
    totalToLandfield: number;
    totalWasteRecycled: number;
}